requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
selenium==4.15.2
beautifulsoup4==4.12.2
httpx==0.25.2
webdriver-manager==4.0.1